+++ src/random_variables.rs
@@ -23,11 +23,9 @@ impl_random_variable_for_int!(u16);
 impl_random_variable_for_int!(u32);
 impl_random_variable_for_int!(u64);
 impl_random_variable_for_int!(u128);
-impl_random_variable_for_int!(usize);
 
 impl_random_variable_for_int!(i8);
 impl_random_variable_for_int!(i16);
 impl_random_variable_for_int!(i32);
 impl_random_variable_for_int!(i64);
 impl_random_variable_for_int!(i128);
-impl_random_variable_for_int!(isize);